Fix bug where approx_pos would not be decremented because it was float

+++ include/learned_merge.h
@@ -77,8 +77,8 @@ private:
     int cluster_length = 0;
 #endif
     // approx_pos is always a valid position in iterator.
-    float approx_pos = smallest->guessPositionMonotone(second_smallest->key());
-    approx_pos = std::max(approx_pos, (float)smallest->current_pos());
+    uint64_t approx_pos = (uint64_t)smallest->guessPositionMonotone(second_smallest->key());
+    approx_pos = std::max(approx_pos, smallest->current_pos());
     bool is_overshoot = false;
     if (comparator->compare(smallest->peek(approx_pos),
                             second_smallest->key()) > 0) {
